+++ src/vecinos.h
@@ -19,11 +19,11 @@ using namespace std;
 class InterpolacionVecinos: public Interpolador{
 	public:
 		InterpolacionVecinos();
-		InterpolacionVecinos(const vector<int> &y, int valores_a_agregar);
-		void recalcular(const vector<int> &y, int valores_a_agregar);
+		InterpolacionVecinos(const vector<double> &y, int valores_a_agregar);
+		void recalcular(const vector<double> &y, int valores_a_agregar);
 		double evaluar(double x);
 	private:
-		vector<int> datos_generados;
+		vector<double> datos_generados;
 		int valores_a_agregar;
 		int cant_datos_originales;
 };
